Move app into main project
